# Store expansion locations separately from tokens