--- a/src/aro/Preprocessor.zig
+++ b/src/aro/Preprocessor.zig
@@ -255,6 +255,14 @@ pub fn deinit(pp: *Preprocessor) void {
     pp.expansion_entries.deinit(pp.gpa);
 }
 
+/// Free buffers that are not needed after preprocessing
+fn clearBuffers(pp: *Preprocessor) void {
+    pp.token_buf.clearAndFree();
+    pp.char_buf.clearAndFree();
+    pp.top_expansion_buf.clearAndFree();
+    pp.hideset.clearAndFree();
+}
+
 pub fn expansionSlice(pp: *Preprocessor, tok: Tree.TokenIndex) []Source.Location {
     const S = struct {
         fn order_token_index(context: void, lhs: Tree.TokenIndex, rhs: Tree.TokenIndex) std.math.Order {
@@ -283,6 +291,7 @@ pub fn preprocessSources(pp: *Preprocessor, sources: []const Source) Error!void 
     try pp.addIncludeResume(first.id, 0, 1);
     const eof = try pp.preprocess(first);
     try pp.addToken(eof);
+    pp.clearBuffers();
 }
 
 /// Preprocess a source file, returns eof token.