@ehaas 2024-02-09
Updated benchmark with `clearBuffers`

```
Benchmark 1 (48 runs): source/arocc/zig-out/bin/arocc -Wno-unreachable-code -fsyntax-only -I. sqlite3.c
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           212ms ±  729us     211ms …  215ms          2 ( 4%)        0%
  peak_rss           62.5MB ± 53.6KB    62.4MB … 62.6MB          0 ( 0%)        0%
  cpu_cycles          788M  ± 1.73M      785M  …  793M           1 ( 2%)        0%
  instructions       1.71G  ± 1.87      1.71G  … 1.71G           0 ( 0%)        0%
  cache_references   2.05M  ± 23.9K     2.02M  … 2.10M           0 ( 0%)        0%
  cache_misses        545K  ± 8.88K      524K  …  575K           3 ( 6%)        0%
  branch_misses      7.34M  ± 14.8K     7.32M  … 7.40M           3 ( 6%)        0%
Benchmark 2 (47 runs): source/expansion-tokens/zig-out/bin/arocc -Wno-unreachable-code -fsyntax-only -I. sqlite3.c
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           213ms ±  642us     212ms …  215ms          1 ( 2%)          +  0.2% ±  0.1%
  peak_rss           58.2MB ± 49.3KB    58.2MB … 58.3MB          0 ( 0%)        ⚡-  6.8% ±  0.0%
  cpu_cycles          795M  ± 2.07M      792M  …  804M           2 ( 4%)          +  0.8% ±  0.1%
  instructions       1.71G  ± 1.79      1.71G  … 1.71G           1 ( 2%)          +  0.2% ±  0.0%
  cache_references   2.02M  ± 21.8K     1.97M  … 2.06M           0 ( 0%)        ⚡-  1.5% ±  0.5%
  cache_misses        547K  ± 9.08K      529K  …  575K           2 ( 4%)          +  0.4% ±  0.7%
  branch_misses      7.25M  ± 9.97K     7.23M  … 7.29M           1 ( 2%)        ⚡-  1.2% ±  0.1%
```

zig2.c: Maximum resident set size (kbytes): `1093240` (previously `1195260`) so saves another 100 megs, not bad.